% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulations.R
\name{create.sim.data}
\alias{create.sim.data}
\title{simulations}
\usage{
create.sim.data(
  sim = 1,
  te = 2,
  jitter = 1,
  cov_min = 0,
  cov_max = 5,
  wt_X1 = 1,
  wt_X2 = 1,
  n = 200,
  mean1 = 0,
  mean2 = 0,
  sd1 = 1,
  sd2 = 1,
  rho = 0.2,
  weight_t1 = -1,
  weight_t2 = 1,
  weight_y0 = 5,
  weight_y1 = -1,
  weight_y2 = 1,
  relX1 = "Mediator",
  relX2 = "Confounder"
)
}
\arguments{
\item{sim}{a integer referencing the simulation - 1, 2, 3, or 4}

\item{te}{a number representing the treatment effect, specify for all simulations.}

\item{jitter}{controls the SD of the rnorm function that is used to add jitter to the outcome variable}

\item{cov_min}{Simulation 2 - minimum value for uniform distribution}

\item{cov_max}{Simulation 2 - maximum value for uniform distribution}

\item{wt_X1}{Simulation 2 - the effect of the treatment on X1}

\item{wt_X2}{Simulation 2 - the effect of the treatment on X2}

\item{n}{the number of observations to populate for simulation 2, 3, and 4}

\item{mean1}{Simulation 3 and 4 - X1 mean value for a normal distribution}

\item{mean2}{Simulation 3 and 4 - X2 mean value for a normal distribution}

\item{sd1}{Simulation 3 and 4 - X1 standard deviation value for a normal distribution}

\item{sd2}{Simulation 3 and 4 - X2 standard deviation value for a normal distribution}

\item{rho}{Simulation 3 and 4 - correlation coefficient for relationship between X1 and X2}

\item{weight_t1}{Simulation 3 and 4 - the weight of X1 on the treatment variable t}

\item{weight_t2}{Simulation 3 and 4 - the weight of X2 on the treatment variable t}

\item{weight_y0}{Simulation 3 and 4 - the unaffected value of the outcome y}

\item{weight_y1}{Simulation 3 and 4 - the weight of X1 on the outcome variable y}

\item{weight_y2}{Simulation 3 and 4 - the weight of X2 on the outcome variable y}

\item{relX1}{Simulation 3 - the causal relationship of X1 in the data. Can be 'mediator', 'confounder', 'collider', 'ancestor to y', 'ancestor to t'}

\item{relX2}{Simulation 3 - the causal relationship of X2 in the data. Can be 'mediator', 'confounder', 'collider', 'ancestor to y', 'ancestor to t'}
}
\value{
A dataframe with a 2 covariates, X1 and X2, a random error variable, a treatment variable and an outcome variable for the purposes of matching
}
\description{
A function for simulating data
}
\examples{
# generate a simulated dataframe
d <- create.sim.data(1,2)
head(d)
}
