% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_matches.R
\name{matched.data}
\alias{matched.data}
\title{matches}
\usage{
matched.data(f, data, dist, order = "data", replace = FALSE)
}
\arguments{
\item{f}{a two sided formula containing the treatment variable on the left side and the matching variables on the right}

\item{data}{a dataframe of the data that is to be used for matching}

\item{dist}{a matching distance. Either "propensity" or "mahalanobis"}

\item{order}{a named order of "data", "smallest", "largest", "random" to perform matches and is only relevant if replace = TRUE
*"data" just uses the order the data is already in
*"smallest" matches by the smallest propensity score first
*"largest" matches by the largest propensity score first
*"random" matches on a random order every time}

\item{replace}{boolean value of whether to replace matched units or not}
}
\value{
Returns a dataframe with the variable t, Allocation, X1, X2, y

A list of items that are generated from the mathcing process
}
\description{
A function for simulating data
}
\examples{
#simulate some data
d <- create.sim.data(1,2)
#matching formula
f <- t~X1+X2
#create matched data object
M <- matched.data(f,d,"Propensity Score")

}
