% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_matchplot.R
\name{matching.plot}
\alias{matching.plot}
\title{Plot and visualise matches from [matched.data()]}
\usage{
matching.plot(match.data, xvar, yvar, multi = F, plot.n = 1)
}
\arguments{
\item{match.data}{An object returned from the matched.data function}

\item{xvar}{A string naming the variable to go on the X axis}

\item{yvar}{A string naming the variable to go on the Y axis}

\item{multi}{Boolean value of whether the output will be part of a plotly subplot}

\item{plot.n}{a number indicating the number plot it is for the title}
}
\value{
A plotly object of the sequence of matches performed on the data
}
\description{
A fct function
}
\examples{
#simulate some data
d <- create.sim.data(1,2)
#matching formula
f <- t~X1+X2
#create matched data object
M <- matched.data(f,d,"Propensity Score")

#create an animated plot of the matched data
matching.plot(M, "X1", "X2", FALSE)

}
