% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulations.R
\name{create.sim.data}
\alias{create.sim.data}
\title{simulations}
\usage{
create.sim.data(
  sim = 1,
  te = 2,
  jitter = 0.5,
  g1_min = 0,
  g1_max = 5,
  g2_shift_X1 = 1,
  g2_shift_X2 = 1,
  n = 200,
  mean1 = 0,
  mean2 = 0,
  sd1 = 1,
  sd2 = 1,
  rho = 0.3,
  weight_t1 = 0.5,
  weight_t2 = 0.5,
  weight_y0 = 2,
  weight_y1 = 1,
  weight_y2 = -1
)
}
\arguments{
\item{sim}{a integer referencing the simulation - 1, 2 or 3}

\item{te}{a number represnting the treatment effect, specify for all simulations.}

\item{jitter}{controls the SD of the rnorm function that is used to add jitter to the outcome variable}

\item{g1_min}{Simulation 2 - minimum value of for uniform distribution}

\item{g1_max}{Simulation 2 - maximum value of for uniform distribution}

\item{g2_shift_X1}{Simulation 2 - how far to shift the X1 value for a control group}

\item{g2_shift_X2}{Simulation 2 - how far to shift the X2 value for a control group}

\item{n}{the number of observations to populate for simulation 2 and 3}

\item{mean1}{Simulation 3 - X1 mean value for a normal distribution}

\item{mean2}{Simulation 3 - X2 mean value for a normal distribution}

\item{sd1}{Simulation 3 - X1 standard deviation value for a normal distribution}

\item{sd2}{Simulation 3 - X2 standard deviation value for a normal distribution}

\item{rho}{Simulation 3 - correlation coefficient for relationship between X1 and X2}

\item{weight_t1}{Simulation 3 - the weight of X1 on the treatment variable t}

\item{weight_t2}{Simulation 3 - the weight of X2 on the treatment variable t}

\item{weight_y0}{Simulation 3 - the unaffected value of the outcome y}

\item{weight_y1}{Simulation 3 - the weight of X1 on the outcome variable y}

\item{weight_y2}{Simulation 3 - the weight of X2 on the outcome variable y}
}
\value{
A dataframe with a 2 covariates, X1 and X2, and a treatment and outcome variable for the purposes of matching
}
\description{
A function for simulating data
}
\examples{
# generate a simulated dataframe
create.sim.data(1,2)
}
