% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_matches.R
\name{matched.data}
\alias{matched.data}
\title{Generate outputs from matching and propensity score methods.}
\usage{
matched.data(f, data, dist, order = "data", replace = FALSE)
}
\arguments{
\item{f}{a two sided formula containing the treatment variable on the left side and the matching variables on the right}

\item{data}{a dataframe of the data that is to be used for matching}

\item{dist}{a matching distance, either "Propensity Score" or "Mahalanobis"}

\item{order}{a named order of "data", "smallest", "largest", "random" to perform matches and is only relevant if replace = TRUE
\itemize{
\item "data" just uses the order the data is already in
\item "smallest" matches by the smallest propensity score first
\item "largest" matches by the largest propensity score first
\item "random" matches on a random order every time
}}

\item{replace}{boolean value of whether to replace matched units or not}
}
\value{
A list of items that are generated from the matching process
}
\description{
A function for simulating data
}
\details{
The output is a list containing the following:
\itemize{
\item \code{paired.data} a dataframe of the matched pairs of data (prefix t. for treated group and c. for control group) with the calculated distance and order of matching
\item \code{matched.data} a dataframe with the matched data ready for analysis
\item \code{data} a dataframe of the original data
\item \code{formula} the formula used for matching
\item \code{distance} the distance used for matching
\item \code{propensity} a numeric vector of propensity scores in the same order as the original data
\item \code{stratification} a factored vector which has stratified the propensity score into groups (between 5 and 10)
\item \code{wt.ATE} a numeric vector of the average treatment effect based on the propensity score
\item \code{wt.ATT} a numeric vector of the average treatment effect on the treated based on the propensity score
\item \code{wt.Matched} a vector of weights based on how the matching occurred
\item \code{d.m} the 'DenseMatrix' output from the \code{\link[optmatch:match_on-methods]{optmatch::match_on()}} function used for matching
\item \code{pairs} a dataframe with the row labels for each treated unit and it's matched control with the calculated distance between them
\item \code{treatment} the treatment variable
\item \code{order} the order the matches were conducted in
\item \code{replacement} was replacement used to match, TRUE or FALSE
}
}
\examples{
#simulate some data
d <- create.sim.data(1,2)
#matching formula
f <- t~X1+X2
#create matched data object
M <- matched.data(f,d,"Propensity Score")

}
